fixed dynamo db table initalization and added initilization to init_database.sh

+++ internal/flush_dynamo_tables.py
@@ -6,12 +6,17 @@ dynamodb = boto3.resource(
     endpoint_url='http://localhost:5600'
 )
 
+def table_exists(table_name):
+    existing_tables = dynamodb.meta.client.list_tables()['TableNames']
+    return table_name in existing_tables
+
 def delete_table(table_name):
     try:
-        table = dynamodb.Table(table_name)
-        table.delete()
+        if table_exists(table_name):
+            table = dynamodb.Table(table_name)
+            table.delete()
     except Exception as e:
-        print(f"Error deleting table '{table_name}':  {e}")
+        print(f"Error deleting table '{table_name}': {e}")
 
 # Delete all tables
 delete_table("User")
@@ -19,5 +24,6 @@ delete_table("Department")
 delete_table("Course")
 delete_table("Enrollment")
 delete_table("Section")
+delete_table("Waitlist")
 
-print("All tables deleted successfully.")
+print("Flushed existing data from dynamodb.")
